agregando userArea
